software_builder/requirements.txt
# Core framework
pydantic==2.8
pyyaml==6.0.2
python-dotenv==1.0.1

# LLM providers
openai==1.45.0
anthropic==0.24.0

# Vector stores
chromadb==0.5.0
pinecone-client==5.0.0

# HTTP utilities
httpx==0.27.2
tenacity==9.0.0   # for retry/back-off logic

# Optional dev / testing tools (feel free to remove if not used)
pytest==8.3.2
black==24.4.2
isort==5.13.2
flake8==7.1.1
